requirements.txt
diffusers>=0.37.0,<0.38.0
transformers>=4.30.0,<5.0.0
accelerate>=1.13.0,<2.0.0
safetensors>=0.7.0,<1.0.0
Pillow>=9.5.0,<11.0.0
numpy>=1.26.2,<3.0.0
huggingface_hub>=0.26.0,<1.0.0
tqdm>=4.67.1,<5.0.0
